pyproject.toml
[tool.poetry]
name = "script"
version = "0.1.0"
description = ""
authors = ["Piotr Bator <piotr.bator.contact@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
openai = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
